requirements.txt
aws-cdk.aws-lambda
aws-cdk-aws-lambda-event-sources
aws-cdk.core